Final changes
Memories are added from Block Memory Genertor with Vivado instead of registers.
A Matlab code is added to compare hardware results with the Matlab Model.
Cordic algorithm has been added to the first stage in order to get rid of registers that coefficients are stored and complex multipliers.
Dedicated Multipliers are added to the 9th and 10th stage.